+++ src/lib.rs
@@ -1,6 +1,5 @@
 pub mod api;
 pub mod compute;
 pub mod cu29;
-pub mod cu29_tasks;
 pub mod pipeline;
 pub mod stats;
